+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Added
 
-- Action to show a diagram based on distributed traces
+- [#37](https://github.com/rafaelldi/aspire-plugin/issues/37): Action to show a diagram based on distributed traces
 
 ### Changed